fastapi/requirements.txt
alembic==1.13.0
asgi_lifespan==2.1.0
celery==5.3.6
diffusers==0.24.0
dramatiq==1.15.0
fastlid==0.1.9
httpx==0.26.0
httpx_ws==0.4.3
langchain==0.0.353
langchain_community==0.0.7
langchain_core==0.1.4
loguru==0.7.2
matplotlib==3.8.2
motor==3.3.2
numpy==1.26.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pydantic_settings==2.1.0
pymongo==4.6.1
pytest==7.4.3
pytest_asyncio==0.23.2
python_jose==3.3.0
Requests==2.31.0
safetensors==0.4.1
scikit_learn==1.3.2
setuptools==69.0.2
SQLAlchemy==2.0.23
starlette==0.34.0
tensorflow==2.14.0
tensorflow_intel==2.14.0
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
ultralytics==8.0.227
uvicorn==0.25.0
